requirements.txt
aiogram==3.10.0
python-dotenv==0.21.1
google-generativeai==0.7.2
llama-cpp-python==0.2.84
watchdog==4.0.1
